LookupService.handleGetStockOptionByMinExpirationAndMinStrikeDistance: use correct lookup method
StockOptionDao.findByMinExpiration... use BigDecimal strike param instead of double

git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@4959 19eef36c-86f0-de11-afab-0004ac968479

+++ core/src/main/java/com/algoTrader/service/StockOptionServiceImpl.java
@@ -232,8 +232,8 @@ public class StockOptionServiceImpl extends StockOptionServiceBase {
             return null;
         }
 
-        List<StockOption> callOptions = getStockOptionDao().findByMinExpirationAndStrikeLimit(1, 1, underlying.getId(), date, underlyingTick.getLast().doubleValue(), OptionType.CALL);
-        List<StockOption> putOptions = getStockOptionDao().findByMinExpirationAndStrikeLimit(1, 1, underlying.getId(), date, underlyingTick.getLast().doubleValue(), OptionType.PUT);
+        List<StockOption> callOptions = getStockOptionDao().findByMinExpirationAndStrikeLimit(1, 1, underlying.getId(), date, underlyingTick.getLast(), OptionType.CALL);
+        List<StockOption> putOptions = getStockOptionDao().findByMinExpirationAndStrikeLimit(1, 1, underlying.getId(), date, underlyingTick.getLast(), OptionType.PUT);
 
         StockOption callOption = null;
         if (!callOptions.isEmpty()) {
